Untested incomplete code

+++ implicitkernel/include/viewer.h
@@ -94,7 +94,14 @@ namespace viewer
     void set_work_group_size();
     static void pause_render_loop();
     static void resume_render_loop();
-    static void add_render_data(uint8_t* bytes, size_t nBytes, uint8_t* types, uint32_t* offsets, size_t nEntities, op_step* steps, size_t nSteps);
+    static void add_render_data(
+        uint8_t* bytes,
+        size_t nBytes,
+        uint32_t* types,
+        uint32_t* offsets,
+        size_t nEntities,
+        op_step* steps,
+        size_t nSteps);
 
     void show_entity(entities::ent_ref entity);
